Move components to their own files
