+++ src/actual-pages/InstructorsPage/InstructorsPage.tsx
@@ -1,5 +1,7 @@
 import { Instructor } from "@/types";
-import { FormEvent, useEffect, useState } from "react";
+import { useEffect, useState } from "react";
+import { CreateForm } from "./CreateForm";
+import { SaveDeleteForm } from "./SaveDeleteForm";
 
 export function InstructorsPage() {
   const [instructors, setInstructors] = useState<Instructor[] | null>(null);
@@ -60,55 +62,5 @@ export function InstructorsPage() {
       ))}
       <CreateForm onCreate={handleCreate} />
     </>
-  );
-}
-
-type CreateFormProps = {
-  onCreate: (name: string) => Promise<void>;
-};
-
-function CreateForm({ onCreate }: CreateFormProps) {
-  const [name, setName] = useState("");
-
-  const handleCreate = async (e: FormEvent<HTMLFormElement>) => {
-    e.preventDefault();
-    await onCreate(name);
-    setName("");
-  };
-
-  return (
-    <form onSubmit={handleCreate}>
-      <input value={name} onChange={(e) => setName(e.target.value)} />
-      <button type="submit">Create</button>
-    </form>
-  );
-}
-
-type SaveDeleteFormProps = {
-  instructor: Instructor;
-  onSave: (id: number, name: string) => Promise<void>;
-  onDelete: (id: number) => Promise<void>;
-};
-
-function SaveDeleteForm({ instructor, onSave, onDelete }: SaveDeleteFormProps) {
-  const [name, setName] = useState(instructor.name);
-
-  const handleSave = async (e: FormEvent<HTMLFormElement>) => {
-    e.preventDefault();
-    await onSave(instructor.id, name);
-  };
-
-  const handleDelete = async () => {
-    await onDelete(instructor.id);
-  };
-
-  return (
-    <form onSubmit={handleSave}>
-      <input value={name} onChange={(e) => setName(e.target.value)} />
-      <button type="submit">Save</button>
-      <button type="button" onClick={handleDelete}>
-        Delete
-      </button>
-    </form>
   );
 }
